+++ code/src/components/SecondQuestion.js
@@ -6,13 +6,14 @@ const SecondQuestion = ({
   onPreviousQuestionChange,
   onNextQuestionChange,
 }) => {
-  const ageGroup = ["3 ", "7", "10", "more than 15"]
+  const ageGroup = ["3 days ", "7 days ", "2 weeks", "1 month"]
 
   return (
     <form id="secondQuestion">
-      <p className="question">How many days would you like to stay?</p>
+      <p className="question">How long would you like to stay there?</p>
+      {/* <div> */}
       {ageGroup.map((group) => (
-        <label className="label" key={group}>
+        <label className="radio-label" key={group}>
           <input
             className="input"
             type="radio"
@@ -24,17 +25,20 @@ const SecondQuestion = ({
           {group}
         </label>
       ))}
+      {/* </div> */}
 
       <div className="button-container">
         <button className="glow-on-hover" onClick={onPreviousQuestionChange}>
           PREVIOUS
         </button>
         <button className="glow-on-hover" onClick={onNextQuestionChange}>
-          NEXT
+          CONTINUE
         </button>
       </div>
     </form>
   )
 }
+
+/* <button onClick={() => window.location.reload(false)} className="btn">Let' start over! <span role="img" aria-label="handpointer">👋🏼</span></button> */
 
 export default SecondQuestion
